Page query is done in CmsPage instead of CmsPageContent.
PT-1126.
