+++ src/pages/cms-page/[pageId].tsx
@@ -2,6 +2,7 @@
 import { NormalizedCacheObject } from '@apollo/client';
 import { GetStaticPropsContext, GetStaticPropsResult, NextPage } from 'next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import { useRouter } from 'next/router';
 import React from 'react';
 
 import { ALL_I18N_NAMESPACES, SUPPORTED_LANGUAGES } from '../../constants';
@@ -16,15 +17,38 @@ import {
   MenuQueryVariables,
   PageIdType,
   Page,
+  usePageQuery,
 } from '../../generated/graphql-cms';
 import { createCmsApolloClient } from '../../headless-cms/cmsApolloClient';
+import { useCMSClient } from '../../headless-cms/cmsApolloContext';
 import CmsPageContent, {
   getUriID,
 } from '../../headless-cms/components/CmsPageContent';
 import { MENU_NAME } from '../../headless-cms/constants';
+import useLocale from '../../hooks/useLocale';
 import { Language } from '../../types';
 
-const CmsPage: NextPage = () => <CmsPageContent />;
+const CmsPage: NextPage = () => {
+  const {
+    query: { pageId },
+  } = useRouter();
+  const locale = useLocale();
+  const cmsClient = useCMSClient();
+
+  const { data: pageData } = usePageQuery({
+    client: cmsClient,
+    variables: {
+      id: getUriID(pageId as string, locale),
+      idType: PageIdType.Uri,
+    },
+  });
+
+  return (
+    <div>
+      <CmsPageContent page={pageData?.page as Page} />;
+    </div>
+  );
+};
 
 export async function getStaticPaths() {
   const cmsClient = createCmsApolloClient();
